feat: Improve djbr extractor
